# Food-Delivery/api_gateway_service/docs/swagger.yaml
basePath: /
definitions:
...
  notification.CreateUserNotificationRequest:
    properties:
      message:
        description: Bildirishnoma matni
        type: string
      order_id:
        description: Bildirishnomaga bog'langan buyurtma ID'si
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.NotificationType'
        description: Bildirishnoma turi
      user_id:
        description: Bildirishnomani oladigan foydalanuvchi ID'si
        type: string
    type: object
  notification.NotificationType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      NotificationType_ORDER_CONFIRMATION: Buyurtmani tasdiqlash bildirishnomasi
      NotificationType_STATUS_UPDATE_USER: Holatni yangilash bildirishnomasi
    x-enum-varnames:
    - NotificationType_ORDER_CONFIRMATION
    - NotificationType_STATUS_UPDATE_USER
...
  notification.UserNotification:
    properties:
      created_at:
        description: Bildirishnoma yaratilgan vaqt
        type: string
      is_read:
        description: Foydalanuvchi bildirishnomani o'qiganligini bildiradi
        type: boolean
      message:
        description: Bildirishnoma matni
        type: string
      notification_id:
        description: Bildirishnoma ID'si
        type: string
      order_id:
        description: Bildirishnomaga bog'langan buyurtma ID'si
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.NotificationType'
        description: Bildirishnoma turi
      user_id:
        description: Bildirishnomani oladigan foydalanuvchi ID'si
        type: string
    type: object
  notification.UserNotificationListResponse:
    properties:
      notifications:
        description: Foydalanuvchi bildirishnomalari ro'yxati
        items:
          $ref: '#/definitions/notification.UserNotification'
        type: array
    type: object
  notification.UserNotificationRequest:
    properties:
      notification_id:
        description: Bildirishnoma ID'si
        type: string
    type: object
  notification.UserNotificationResponse:
    properties:
      message:
        description: Javob xabari
        type: string
      notification:
        allOf:
        - $ref: '#/definitions/notification.UserNotification'
        description: Foydalanuvchi bildirishnomasi ma'lumotlari (Create operatsiyasi
          uchun)
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for Food Delivery System resources
  title: Food Delivery System API
  version: "1.0"
paths:
...
  /api/notifications/user:
    get:
      description: Retrieve a list of all user notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.UserNotificationListResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all user notifications
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Add a new user notification to the system
      parameters:
      - description: Notification Data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification.CreateUserNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification.CreateUserNotificationRequest'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new user notification
      tags:
      - Notifications
  /api/notifications/user/{notificationId}:
    delete:
      description: Remove a user notification from the system by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.UserNotificationResponse'
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid Notification ID
          schema:
            type: string
        "404":
          description: Notification Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a user notification
      tags:
      - Notifications
    get:
      description: Retrieve details of a user notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.UserNotificationRequest'
        "400":
          description: Invalid Notification ID
          schema:
            type: string
        "404":
          description: Notification Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user notification by ID
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"